Add the ability to change the Tika Server's log file path
Added a variable to allow the log file for the Tika server to be
different to that of the Jar file path. This is valuable when changing
the TikaJarPath as well. In my case I am putting the Jar file in a
permanent location other than /tmp, but I want the log file to be in
/var/log/ where it can be rotated automatically.